+++ packages/blade/src/components/Alert/Alert.tsx
@@ -2,14 +2,28 @@ import type { ReactChild, ReactElement } from 'react';
 import { Fragment, useState } from 'react';
 
 import { StyledAlert } from './StyledAlert';
-import { AlertOctagonIcon, AlertTriangleIcon, CheckCircleIcon, CloseIcon } from '~components/Icons';
-import { getPlatformType, makeAccessible } from '~utils';
+import {
+  AlertOctagonIcon,
+  AlertTriangleIcon,
+  CheckCircleIcon,
+  CloseIcon,
+  InfoIcon,
+} from '~components/Icons';
+import {
+  getPlatformType,
+  makeAccessible,
+  metaAttribute,
+  MetaConstants,
+  useBreakpoint,
+} from '~utils';
 import { IconButton } from '~components/Button/IconButton';
 import Box from '~components/Box';
 import { Heading, Text } from '~components/Typography';
 import BaseButton from '~components/Button/BaseButton';
 import { BaseLink } from '~components/Link/BaseLink';
 import type { ColorContrastTypes, Feedback } from '~tokens/theme/theme';
+import { useTheme } from '~components/BladeProvider';
+import type { DotNotationSpacingStringToken } from '~src/_helpers/types';
 
 type Nullable<Type> = Type | null;
 
@@ -76,9 +90,9 @@ type AlertProps = {
   /**
    * Sets the color tone
    *
-   * @default information
+   * @default neutral
    */
-  intent?: Exclude<Feedback, 'neutral'>;
+  intent?: Feedback;
 
   /**
    * Removes border and border radii, useful for creating full bleed layouts. Automatically sets `isFullWidth` to `true` when enabled.
@@ -112,7 +126,8 @@ const CloseButtonWrapper = isReactNative ? Box : Fragment;
 const intentIconMap = {
   positive: CheckCircleIcon,
   negative: AlertOctagonIcon,
-  information: AlertOctagonIcon,
+  information: InfoIcon,
+  neutral: InfoIcon,
   notice: AlertTriangleIcon,
 };
 
@@ -123,7 +138,7 @@ const Alert = ({
   onDismiss,
   contrast = 'low',
   isFullWidth = false,
-  intent = 'information',
+  intent = 'neutral',
   isBorderless = false,
   actions,
 }: AlertProps): ReactElement | null => {
@@ -132,25 +147,71 @@ const Alert = ({
       '[Blade: Alert]: SecondaryAction is allowed only when PrimaryAction is defined.',
     );
   }
+  const { theme } = useTheme();
+  const { matchedDeviceType } = useBreakpoint({ breakpoints: theme.breakpoints });
+  const isMobile = matchedDeviceType === 'mobile';
+
   const [isVisible, setIsVisible] = useState(true);
   const contrastType = `${contrast}Contrast` as const;
+  const iconSize = isBorderless ? 'large' : 'medium';
+  const textSize = isBorderless ? 'medium' : 'small';
 
   const Icon = intentIconMap[intent];
-  const icon = <Icon color={`feedback.icon.${intent}.${contrastType}`} size="large" />;
+  let iconOffset: DotNotationSpacingStringToken = 'spacing.1';
+
+  // certain special cases below needs special care for near perfect alignment
+  if (isReactNative) {
+    if (isBorderless && !title) {
+      iconOffset = 'spacing.1';
+    } else if (!isBorderless && !title) {
+      iconOffset = 'spacing.0';
+    } else if (!isBorderless && title) {
+      iconOffset = 'spacing.2';
+    }
+  } else if (isMobile) {
+    if (!isBorderless && title) {
+      iconOffset = 'spacing.2';
+    } else if (isBorderless && !title) {
+      iconOffset = 'spacing.2';
+    }
+  }
+
+  const icon = (
+    <Box marginTop={iconOffset} display="flex">
+      <Icon color={`feedback.icon.${intent}.${contrastType}`} size={iconSize} />
+    </Box>
+  );
 
   const _title = title ? (
     <Box marginBottom="spacing.2">
-      <Heading size="small" contrast={contrast}>
-        {title}
-      </Heading>
+      {isBorderless ? (
+        <Heading size="small" contrast={contrast}>
+          {title}
+        </Heading>
+      ) : (
+        <Text weight="bold" contrast={contrast}>
+          {title}
+        </Text>
+      )}
     </Box>
   ) : null;
 
-  const _description = <Text contrast={contrast}>{description}</Text>;
+  const _description = (
+    <Box marginTop={title || isReactNative ? 'spacing.0' : 'spacing.1'}>
+      <Text size={textSize} contrast={contrast}>
+        {description}
+      </Text>
+    </Box>
+  );
 
   const primaryAction = actions?.primary ? (
     <Box marginRight="spacing.5" display={isReactNative ? 'flex' : 'inline-flex'}>
-      <BaseButton onClick={actions.primary.onClick} intent={intent} contrast={contrast}>
+      <BaseButton
+        size={textSize}
+        onClick={actions.primary.onClick}
+        intent={intent}
+        contrast={contrast}
+      >
         {actions.primary.text}
       </BaseButton>
     </Box>
@@ -175,7 +236,7 @@ const Alert = ({
   }
   const secondaryAction = actions?.secondary ? (
     <SecondaryActionWrapper>
-      <BaseLink contrast={contrast} intent={intent} {...secondaryActionParams}>
+      <BaseLink size={textSize} contrast={contrast} intent={intent} {...secondaryActionParams}>
         {actions.secondary.text}
       </BaseLink>
     </SecondaryActionWrapper>
@@ -201,7 +262,7 @@ const Alert = ({
         accessibilityLabel="Dismiss alert"
         onClick={onClickDismiss}
         contrast={contrast}
-        size="large"
+        size={iconSize}
         icon={CloseIcon}
       />
     </CloseButtonWrapper>
@@ -225,9 +286,10 @@ const Alert = ({
       isFullWidth={isFullWidth}
       isBorderless={isBorderless}
       {...a11yProps}
+      {...metaAttribute(MetaConstants.Component, MetaConstants.Alert)}
     >
       {icon}
-      <Box flex={1} paddingLeft="spacing.4" paddingRight="spacing.2">
+      <Box flex={1} paddingLeft={isBorderless ? 'spacing.4' : 'spacing.3'} paddingRight="spacing.2">
         {_title}
         {_description}
         {_actions}