+++ packages/blade/src/components/Badge/getStyledBadgeStyles.ts
@@ -1,16 +1,14 @@
 import type { CSSObject } from 'styled-components';
 import type { StyledBadgeProps } from './types';
-import { maxWidth, minHeight } from './badgeTokens';
+import { maxWidth } from './badgeTokens';
 import { getIn, makeBorderSize, makeSize } from '~utils';
 
 const getStyledBadgeStyles = ({
   theme,
-  size,
   platform,
   backgroundColor,
 }: StyledBadgeProps): CSSObject => ({
   backgroundColor: getIn(theme.colors, backgroundColor),
-  minHeight: makeSize(minHeight[size]),
   borderRadius: makeBorderSize(theme.border.radius.max),
   maxWidth: makeSize(maxWidth[platform]),
   display: 'flex',